feat(cache): try to fix cache api
Added a "caches" type that you can only access from the worker global
scope in worker-sys so that hopefully wasm-bindgen will understand
what i want.

The abstractions in the higher level crate are for "Cache", so you never
interact with the JS global directly, and instead instantiate a cache
either by name (with a String) or with the `Default` trait, which returns
the global default cache.