fix: responses occasionally corrupted
Fixes issue where responses can have their bodies corrupted. This is
caused by us giving JS the body's bytes as a reference to a byte slice,
which is then dropped causing a use after free. Most of the time this is
fine, but occasionally the allocator will alocate something that
overlaps with the response body.

To fix this, we instead give the Response constructor an owned
Uint8Array that doesn't have any references into the wasm memory.