fix: uninitialized wasm imports
Previously we generated a "export_wasm.mjs" that imported all of the
glue generated by wasm bindgen and modified the glue to import from the
generated file. This caused a circular import leading to some functions,
specifically ones that were just const aliases of builtins, generated by
wasm-bindgen to be uninitialized and cause a crash.

To fix this we move all of the code in the "export_wasm.mjs" to that
glue file.