chore: fix unused function warning
With the addition of integration testing for websockets we
unintentionally introduced a warning where some functions declared in
the integration test's util module were unused. This is because all
integration tests are compiled as seperate binaries so for the websocket
binary functions weren't used for that specific compile job.

This also adjust flags in the CI to catch warnings like this in the
future.