refactor: use esbuild instead of swc for worker bundling
Previously we used a combination of js modification and SWC bundling to
build our workers built with wasm-bindgen but we were relying under an
assumption that the wasm module was never used until it was called into,
which is not the case. The best way to fix this is to have a loader for
the WASM file as wasm-bindgen expects it to be loaded as an instance
when imported but currently swc cannot bundle things with lazy
imports/exports.

So to get around this we're now automatically downloading esbuild if
it's not on the user's path and using that to bundle our workers.