fix: invalid cfg check for fixing file permissions
On unix-like platforms we're required to set the permissions of the
downloaded esbuild binary to have an executable flag. To control these
permissions rust has `std::os::unix::fs::OpenOptionsExt` which is only
available on unix-like targets but because of the way it was handled it
would still try to be imported when compiling for windows, which would
cause an error.