`durable::ScheduledTime`: Properly compute offsets
`From<Duration>` and `From<i64>` for `ScheduledTime` interpret the
inputs as timestamps (in particular, the number of milliseconds since
the UNIX epoch). Based on the test case in worker-sandbox
("durable_alarm"), the intention appears to interpret these as offsets
relative to the current time. This change aligns these methods with this
expectation.