Export worker::cache::CacheKey and worker::env::EnvBinding
The `worker::cache::CacheKey` is useful when the user wants to wrap the
cache methods by accepting the same bounds for the generic key type.

The `worker::env::EnvBinding` export is a follow-up to
https://github.com/cloudflare/workers-rs/pull/318. Without that export,
the user cannot implement the `EnvBinding` trait which is required by
the the `worker::env::Env::get_binding()` method.