Add `response_body` function to `r2::ObjectBody`.
This commit adds a new function to `r2::ObjectBody` to allow
the underlying `ReadableStream` for the body of an R2 object to
be used to create a `ResponseBody`, via the `ResponseBody::Stream`
variant.

This adds support for cases where workers can hand a reference to
the R2 object to the workers runtime for streaming to the client,
unmodified. Importantly, this allows the worker to avoid CPU usage
while streaming the object.

Before this commit, `ObjectBody::Stream` could be used but that
would result in unexpected CPU usage during the streaming process
leading to the working hitting the CPU limit for large (> 150 MB)
R2 objects.

Closes #389